[Fix] Restructure Ascend chunk_gated_delta_rule_bwd_dhu to AICore task loop

+++ fla/modules/backends/triton_ascend/rotary.py
@@ -67,7 +67,7 @@ def rotary_embedding_kernel(
     if IS_VARLEN:
         i_n, i_t = tl.load(chunk_indices + i_t * 2).to(tl.int32), tl.load(chunk_indices + i_t * 2 + 1).to(tl.int32)
         bos, eos = tl.load(cu_seqlens + i_n), tl.load(cu_seqlens + i_n + 1)
-        T = eos - bos
+        T = (eos - bos).to(tl.int32)
         x = x + bos * H*D + i_h * D
         y = y + bos * H*D + i_h * D
     else:
